Fix EventMeshGrpcProducerTest#testPublishWithException
Although EventMeshGrpcProducerTest#testPublishWithException passes,
it's wrong to do so.
This tests calls the EventMeshGrpcProducer#produce method with a valid
EventMeshMessage instance, so it does not, in fact, throw an error.
The test then doesn't fail on the following line. Since no exception
is ever thrown, the catch block is never reached, and the assertion in
it is never called.

This patch fixes the test by passing a truly invalid message (just a
string), and adds an explicit failure if an exception is not called.